Version 1.0
/* #Joel's Floating Pancake Grid
================================================== */
/* Base Grid */

/* Pattern for HTML

Create the rows in your document

<header>
...
</header>
<nav>
...
</nav>

etc.

Between the rows, you will need a continer <div class="container"> ...
</div>
This needs to be in each row, this is what is centered in the site.

The default wordpress navigation system will be converted. Try and match
the number
of columns and main nav items to make it look its best.

If you need columns, add "onecol" "twocol" etc. to the container div
e.g. <div class="container fivecol"> ... </div> In the div, add in the
columns you need

<div class="col">...</div>
<div class="col">...</div>
<div class="col">...</div>
<div class="col">...</div>
<div class="col">...</div>

one for each column. If you want them to span a couple or a few cols,
add spantwo, spanthree etc. to the
col div.

e.g.  <div class="col spanthree">...</div>
*/